fix: video section layout 수정
