+++ assignmentYoutube/assignmentYoutube/Global/UICollectionView+Extension.swift
@@ -1,0 +1,12 @@
+//
+//  UICollectionView+Extension.swift
+//  assignmentYoutube
+//
+//  Created by Roh on 5/17/24.
+//
+
+import UIKit
+
+extension UICollectionViewDataSource {
+//    func isCellEnable(with indexPath: IndexPath, with: )
+}
